+++ snippet_uiautomator/uiautomator.py
@@ -24,6 +24,7 @@ from mobly import utils as mobly_utils
 from mobly.controllers import android_device
 from mobly.controllers.android_device_lib import adb
 from mobly.controllers.android_device_lib.services import base_service
+from mobly.snippet import errors as snippet_errors
 
 from snippet_uiautomator import configurator as uiconfig
 from snippet_uiautomator import errors
@@ -141,7 +142,18 @@ class UiAutomatorService(base_service.BaseService):
 
     if self._configs.snippet.package_name is None:
       raise errors.ConfigurationError(errors.ERROR_WHEN_PACKAGE_NAME_MISSING)
-    self._device.load_snippet(self._service, self._configs.snippet.package_name)
+
+    start_time = utils.get_latest_logcat_timestamp(self._device)
+    try:
+      self._device.load_snippet(
+          self._service, self._configs.snippet.package_name
+      )
+    except snippet_errors.ServerStartProtocolError as e:
+      if utils.is_service_registered(self._device, start_time):
+        raise errors.UiAutomationServiceAlreadyRegisteredError(
+            self._device.serial, errors.ERROR_WHEN_SERVICE_ALREADY_REGISTERED
+        ) from e
+      raise
 
   def _initial_uidevice(self) -> None:
     """Initializes the UiDevice object."""
@@ -246,3 +258,4 @@ def unload_uiautomator_service(ad: android_device.AndroidDevice) -> None:
   """Stops Snippet UiAutomator service."""
   if ad.services.has_service_by_name(ANDROID_SERVICE_NAME):
     ad.services.unregister(ANDROID_SERVICE_NAME)
+